+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - Added Authentication access to intro modal.
 - External links to chrome for website link and streetview.
+- Checking on the pasted URLs.
 
 ## 0.1.10 (2022-09-19)